requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
tensorflow
pillow
numpy
streamlit